requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorboard
tensorflow-gpu
keras
tqdm
bcolz
nltk
gensim
h5py
tables
pillow
xgboost
line_profiler
memory_profiler
missingno
jupyter
jupyter_contrib_nbextensions
jupyter_nbextensions_configurator
jupyterthemes
